As/mission facade tests

+++ Secret-agency-api/src/main/java/cz/fi/muni/pa165/secretagency/dto/MissionCreateDTO.java
@@ -13,10 +13,10 @@ import java.util.Set;
 public class MissionCreateDTO {
 
     @NotNull
-    private Long latitude;
+    private Double latitude;
 
     @NotNull
-    private Long longitude;
+    private Double longitude;
 
     @NotNull
     private MissionTypeEnum missionType;
@@ -28,19 +28,19 @@ public class MissionCreateDTO {
 
     private Set<Long> agentIds;
 
-    public Long getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Long latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public Long getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Long longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
